# internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package sqlc

import (
	"database/sql"

	"github.com/shopspring/decimal"
)

type Expense struct {
	ID          int32
	Name        string
	Description string
	Category    string
	Amount      decimal.Decimal
	CreatedAt   sql.NullTime
	UpdatedAt   sql.NullTime
	UserID      int32
}

type User struct {
	ID       int32
	Email    string
	Password string
	Username string
}